🐛 fix a bug in reading the initial layout from Qiskit `QuantumCircuit`s
the ancilla register that Qiskit uses for denoting the qubits not initially used after mapping to an architecture, was merely ignored in the past.
if a non-trivial initial layout was selected in the mapping, this could result in a scenario where duplicate entries where present in the initial layout whereas some entries were missing.
The fix is to not ignore the ancilla register but rather add it at the very end and mark the corresponding qubits as ancillaries.

Signed-off-by: Lukas Burgholzer <lukas.burgholzer@jku.at>