♻️ Refactor dynamic casts in Simulation.cpp for better readability
Code in multiple places was refactored to check if a given operation in the quantum computing simulation is non-unitary, classic controlled, or a measurement. The dynamic_cast is now performed in the if condition itself to make it more readable. Also, instead of repeating if clauses, we use one if clause to cast it into a non-unitary operation and then use subsequent if conditions to check its type. This change was made to increase the readability and maintainability of the code.