♻️ Refactor the `measureOneCollapsing` method to make it more readable
The `measureOneCollapsing` function was refactored to increase the code readability. The comments were expanded to provide a better understanding of how the function works. A new helper function `performCollapsingMeasurement` was introduced to decrease the complexity of the `measureOneCollapsing` function, provoking its readability. The code was also modify to avoid initializing some local variables without an initial value, thus, removing some linter warnings.