♻️ add some `[[nodiscard]]` and remove redundant `inline`
The commit introduces the [[nodiscard]] attribute to several functions in the ComplexNumbers class. This attribute ensures that the return value of these functions is used or explicitly discarded, which can effectively reduce the potential for some bugs to go unnoticed. It promotes more mindful use of the functions as the compiler will now issue a warning if the return value is not used.