♻️ Add `noexcept` specifier to functions in the Complex and ComplexTable classes.
This change is made to indicate that these functions do not throw any exceptions. This not only improves performance by having less code generated for exception handling, but it also makes it easier for developers to reason about the code by knowing which functions can potentially throw exceptions. All of the modified functions do not throw any exceptions, so they should be marked with noexcept to relay this information to the compiler and enable certain optimizations.