✨ Introduce `RealNumber` class that replaces `CTEntry`
This update lessens introduces a `RealNumber` class to replace `CTEntry`. The motivation for this is many-fold. First, this change disentangles the ComplexTable and the data type it stores (similar to how `Node` is separated from `UniqueTable`). Second, it clears up the misnomer of calling it a _complex_ entry, when it only ever could represent a single real number. Third, it provides a cleaner interface and library structure, e.g., allowing to forward declare the `RealNumber` struct in headers and only include it in source files (hence, improving compilation time).