🔥 Remove ToffoliTable from Package and simplify X gate handling in Operations
Deleted the ToffoliTable from the Package to eliminate unnecessary complexity. Additionally, simplified the handling of the X gate within Operations to further decrease complexity. This was possible as the ToffoliTable was only used in this location. The removal does not lead to a loss of information or functionality, only a simplification of the code structure.

Signed-off-by: Lukas Burgholzer <lukas.burgholzer@jku.at>