🐛 fix compute table lookup
Due to terminals no longer being represented by separate entities, some of the checks whether a compute table lookup was successful did not work anymore as they checked for `nullptr`. This commit adjusts the `lookup` function to return a pointer to an entry. This makes it easier to distinguish a successful lookup from one that failed.

Signed-off-by: Lukas Burgholzer <lukas.burgholzer@jku.at>