♻️ Refactor handling of Barrier operations
This commit refactors how the quantum barriers are handled across the operations, moving it from nonunitary operations to the standard operations. The changes include updates in the display of the barriers and how they are parsed in the QASMParser, along with modifications in the operation files and test files. This change improves code structure allowing a more consistent handling of operations.

Signed-off-by: Lukas Burgholzer <lukas.burgholzer@jku.at>