♻️ Refactor handling of Barrier operations (#394)
## Description

This commit refactors how the `Barrier` operations are handled across
the operations, moving it from non-unitary operations to the standard
operations. The changes include updates in the display of the barriers
and how they are parsed in the QASMParser, along with modifications in
the operation files and test files. This change improves code structure
allowing a more consistent handling of operations. Barriers are
essentially No-Ops, so there is no reason for them to be non-unitary.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

Signed-off-by: Lukas Burgholzer <lukas.burgholzer@jku.at>