♻️ Refactor GitHub Actions and add reusable workflows
A series of smaller reusable workflows have been created. These include workflows for different types of tests (Python, C++, Windows, MacOS), Python packaging, and linting. The purpose of this change is to improve the modularity and readability of the workflows, and allow for better reuse across different workflows and projects.

Furthermore, code coverage reports are now split between Python and C++ test workflows and will be automatically uploaded to Codecov using the flags feature. Carry-forward support is enable for scenarios where not all reports are uploaded.

Lastly, CI and CD workflows were split for better readability and maintainability. Wheels are now only built on releases.

Signed-off-by: burgholzer <burgholzer@me.com>