♻️ Refactor `NonUnitaryOperation` class to simplify target handling (#399)
## Description

In the NonUnitaryOperation class, the previous implementation used two
different vectors to manage the targets for Measure and other
operations, which lead to redundancy in the code. This update unifies
the handling of targets by using only one vector (named targets) for all
operations. This simplification enhances the code readability and
maintainability. It also removes the need of several functions like
getTargets() and getUsedQubits(). Several function argument lists have
also been refactored for cleaner and more consistent usage. Further, the
print() function was restructured for better clarity. Overall, these
changes will be beneficial for future development work.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

Signed-off-by: burgholzer <burgholzer@me.com>