Add invert functionality for quantum operations (#416)
## Description

Add an `invert` function to `Operation`. This mirrors the functionality
of the [OpenQASM 3.0 `inv`
modifier](https://openqasm.com/language/gates.html#inverse-modifier).

For Compound gates, this calls `invert` on each nested operation and
reverses the order of the operations.
For Standard gates, this checks which type of gate is being inverted and
handles each case.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.