♻️ Refactor argument order for methods in QuantumComputation.hpp
Methods have been refactored in QuantumComputation to adjust the ordering of arguments, specifically the target, control, and parameters. The order is now consistent with Qiskit. The reasoning behind this change is to provide consistency and clarity in the way the methods are called. This reduces the friction for someone that knows Qiskit to use the MQT.

Signed-off-by: burgholzer <burgholzer@me.com>