✨♻️ Refactor `QuantumComputation` API and Expose to Python (#371)
To make mqt-core accessible from python - in particular to make commonly used mqt-core methods and datatypes available to other mqt tools in python - this PR aims to port the core functionality to python. 

This PR does not aim to provide every functionality available in the C++ library to python. This is left for future changes when the demand for such functionality arises.

In the process, some of the interfaces of the `QuantumComputation` class where adjusted drastically.
This constitutes a breaking change and will require consuming libraries to adapt.
However, it greatly simplifies some constructions and makes the interface way cleaner.
In particular,
- some operation names have changed to more closely map to the Qiskit names
- the order of function parameters in the `QuantumComputation` class was switched to be more natural and in-line with what other SDKs are providing. The same convenience functions were given more suitable names to indicate their functionality (e.g., `x`, `cx`, and `mcx`). Lots of boilerplate code was reduced and refactored.
- `Control` objects are now implicitly constructible from `Qubit` indices which, again, simplifies a lot of constructs
- Operations and Quantum Computations are now copyable by default. There never was a good reason for them to not be copyable and it limited us and the API in quite some ways. This makes use of the previously defined `clone` methods.
- The reliability of circuit printing was improved considerably.
- The new and native Qiskit import can properly handle Initial Layouts composed of multiple registers in not-necessarily alphabetical order. This should eliminate some error that people encountered in certain applications, e.g., using QCEC.

Last, but certainly not least, this PR also adds full API documentation for the exposed functionality and a Quickstart Guide to get started.

---------

Signed-off-by: burgholzer <burgholzer@me.com>
Co-authored-by: burgholzer <burgholzer@me.com>