♻️ simpler gate matrices (#449)
## Description

A second PR in the series of small improvements. Builds on #448 and
simplifies the construction of gate matrices throughout the library.
Furthermore proper naming is applied so that clang-tidy ignores are not
necessary anymore.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.