⚡ Small unique table lookup improvement (#451)
## Description

The third PR in the little series of improvements from #444. This time a
small logic improvement in the unique table lookup logic. It only ever
needed to work on nodes as it did not consider the top edge weight
anyways.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [ ] I have made sure that all CI jobs on GitHub pass.
- [ ] The pull request introduces no new warnings and follows the
project's style guidelines.