⚡ Completely eliminate temporary numbers borrowed from the cache (#457)
## Description

This PR is "just" another one in the series of bringing over
improvements from #444.
This time it starts with the reduction of the complex cache being used
by completely eliminating any temporary use of cache entries. This is
enabled by extending the computation support and interoperability
between `Complex` and `ComplexValue`.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.