Fix import error.
On my platform (Ubuntu 20.04) there is an error due to the definition of `symbolic` and `operations` as submodules of
`mqt.core` since this causes some classes to be defined multiple times, which crashes the Python interpreter when
importing `mqt.core`. I haven't found another workaround than importing everything in the top module.