:sparkles: added state preparation routines for GHZ and W state (#445)
## Description

<!-- Please include a summary of the change and which issue is fixed.
Please also include relevant motivation and context. List any
dependencies that are required for this change. -->

In
#[341](https://github.com/cda-tum/mqt-core/issues/341#issue-1758878364)
the DD package provided methods for directly constructing DD for
computational basis states or any product state that can be constructed
from the single-qubit states `|0>, |1>, |+>, |->, |L>, |R>`.

## Updates
There are now 2 routines which implement the 
- GHZ state
- W state,
also, there are unit tests which check that the preparation routines
provide the desired states for a smaller number of qubits.

Fixes
#[341](https://github.com/cda-tum/mqt-core/issues/341#issue-1758878364)
<!--- Replace (issue) with the issue number that is fixed by this pull
request. -->

## Notes
In the beginning, I tried adding the routines to a different header
file, however, I noticed that it was a lot easier to stick to the
current structure and add the functions to the Package.hpp header inside
the DD package. Additionally, I tried to use the existing functions in
order to create & normalize the decision diagrams.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Lukas Burgholzer <burgholzer@me.com>