:sparkles: added inverse gate for iSWAP (#460)
## Description

The DD Package was missing the ```iSWAPDag``` (iSWAP dagger) gate.
However, the DD Package already had in some place a ```ISWAPDG_MAT```
which implemented the desired gate in a matrix form.

## Status
Currently, the `iSWAPdagDumpIsValid` test produces the following error:
```
gate type iswapdg could not be converted to OpenQASM

i:   0   1
1:iswdiswd
o:   0   1

~/mqt-core/test/unittests/test_io.cpp:389: Failure
Expected: qc->import(ss, qc::Format::OpenQASM); doesn't throw an exception.
  Actual: it throws std::runtime_error with description "[qasm parser] l:6 c:2 msg: Unknown gate q".


i:   0   1
o:   0   1
```

After reading through
[https://qiskit.org/documentation/apidoc/circuit_library.html](url), my
assumption is that OpenQASM does not support at the moment the iSWAP
dagger gate natively, so, therefore, it doesn't have implementation for
it.

(I know that the previous assumption might not be the reason why, but I
tried to point out a tiny direction why this might not work. Anyway, I
will try to look further into this issue in the near future)

Fixes #423 <!--- Replace (issue) with the issue number that is fixed by
this pull request. -->

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [ ] I have added appropriate tests and documentation.
- [ ] I have made sure that all CI jobs on GitHub pass.
- [ ] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>