🐛 fix performance regression due to premature rounding
Due to the way we approximate values during a computation `( * b) * c != a * (b * c)`. Such a case happened in the multiplication routine when the top edge weight returned by the multiplication or the compute table was large, but the operands top edge weights were small enough so that their product would round to zero.