🎨 provide `operator<<` overloads for scoped enums in ZX package (#508)
## Description

While this may not solve the underlying issue in #503, it should at
least resolve some part of it and maybe reveal further warnings. It also
fixes one of the new clang-tidy warnings for reducing the size of the
underlying data type of enums.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.