🐛 avoid using system header includes for external dependencies (#517)
## Description

This PR changes all includes of external dependencies that are shipped
with the project to use double quotes instead of angle brackets. As a
consequence, this should pick up the shipped versions first instead of
the system ones.

Fixes #503 

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.