✨ Allow project installation (#529)
## Description

This PR sets up the project so that it can be installed properly.
As it is configured right now, this allows to install the `mqt-core`
Python package and make the C++ project available via
`find_package(mqt-core)` as long as the proper cmake directory is added
to the search path.
The right directory can easily be determined via `python -m mqt.core
--cmake_dir`.
You might be wondering what this is useful for. It allows us to only
require a single source copy of mqt-core in top-level projects once we
switch them to using the `mqt-core` Python package. Otherwise, we would
need to maintain the submodule and the Python package in parallel.

While I expect this to require some follow ups once used in practice, I
believe this is a solid start.

Fixes #504 

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [ ] I have added appropriate tests and documentation.
- [ ] I have made sure that all CI jobs on GitHub pass.
- [ ] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Signed-off-by: burgholzer <burgholzer@me.com>