:zap: Added native construction of two-target controlled-gate DDs (#534)
## Description
The purpose of this PR is to add the extra logic for two-target
controlled-gate DDs mentioned in
https://github.com/cda-tum/mqt-core/issues/334#issue-1758497551. The
problem was that the current implementation of the `makeTwoQubitGateDD`
only worked for non-controlled two-target gates. Additionally, what we
also want to achieve with this PR is to get rid of all the extra methods
using the previous workaround
(i.e.https://github.com/cda-tum/dd_package/blob/7c9644a2a65742bb580e3e5fefeac075c9534351/include/dd/Package.hpp#L629-L810).

At the moment the updated version of `makeTwoQubitGateDD` is passing the
CI, but there are still some TODOs left:
- [x] test for equivalence between old-fashioned workaround methods and
`makeTwoQubitGateDD`
- [x] update the code across the package (e.g. getStandardOperationDD)
- [x] verify the functionality by testing for edge cases

<!---Please include a summary of the change and which issue is fixed.
Please also include relevant motivation and context. List any
dependencies that are required for this change.-->

Fixes #334 <!--- Replace (issue) with the issue number that is fixed by
this pull request. -->

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Lukas Burgholzer <burgholzer@me.com>