💥 build `manylinux_2_28` wheels instead of `manylinux2014`
`manylinux2014` forces old C++ ABI instead of the modern `cxx11` ABI. This creates all kinds of problems when trying to compile the libraries shipped with the `mqt.core` package together with code being compiled on a more modern system that defaults to using the new ABI.
Since `manylinux2014` is reaching EOL in 2024, it does not seem too drastic to move to a more modern image.

If any user problems should arise because of this, we might simply ship both kinds of wheels in the future.

Signed-off-by: burgholzer <burgholzer@me.com>