Fixed incorrect handling of global phases.
When floating point values are converted to `PiRational`s, they are constrained to the interval (-pi, pi]. This doesn't
matter during computations in the ZX-library, but when converting from a quantum circuit, this can lead to problems when
dealing with global phases. If global phases have to be introduced while parsing a `QuantumComputation, ' the original value of the phase must be passed along so the global phase is set correctly.