✨ add method for computing the conjugate of a vector DD (#566)
## Description

This PR adds a convenience method for efficiently computing the complex
conjugate of a vector decision diagram, i.e., a new decision diagram,
where every amplitude is complex conjugated. This amounts to traversing
the decision diagram and conjugating each edge weight. As such, it
directly scales with the number of nodes in the decision diagram.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

Signed-off-by: burgholzer <burgholzer@me.com>