:sparkles: :recycle: Adopt UV in CI and refactor reusable workflows (#583)
## Description

This PR updates the CI configuration to use `uv` by Astral (the creators
of `ruff`), which is an extremely fast Python package installer and
resolver, written in Rust. It is sesigned as a drop-in replacement for
common `pip` and `pip-tools` workflows.
See https://github.com/astral-sh/uv

Similar to `ruff`, it is extremely fast and in very active development.
This should speed up any of the Python-based CI runs without much
confirguration overhead.

This PR also uses this opportunity to update the reusable workflows to
incorporate `uv` and a cleaner `nox` setup. Furthermore, it eliminates
the separate `coverage` workflow in favor of just generating coverage
data for all python test runs.
It refactors the `minimums` check to make use of `uv`'s
`--resolution=lowest-direct" feature. This allows us to completely get
rid of the `constraints.txt` file.
It also runs the minimums session on the oldest and newest supported
python version instead of only the oldest. This allows to cover certain
special cases that depend on the python version.
Last, but not least, it makes use of some new features of
https://github.com/hynek/build-and-inspect-python-package to dynamically
generate the build matrices for Python versions.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Signed-off-by: burgholzer <burgholzer@me.com>