:sparkles: Support for Qiskit's final_layout (#592)
## Description

This PR adds support for Qiskit's `final_layout` attribute when
importing circuits from Qiskit. It interprets the `output_permutation`
of an MQT `QuantumComputation` object based on Qiskit's
`final_index_layout()`, which maps the original input circuit's qubit
indices to the final physical qubit indices after routing. This mapping
is the result of applying `routing_permutation()` to the
`initial_layout`. Additionally, it fixes the interpretation of the
output_permutation in `initializeIOMapping()` to align with Qiskit's
final_layout.

Fixes #439


## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Signed-off-by: burgholzer <burgholzer@me.com>
Co-authored-by: burgholzer <burgholzer@me.com>