🩹 make GMP dependency opt-in instead of used-if-found (#608)
## Description

This PR changes the CMake configuration so that GMP, which can be used
in the ZX library, needs to be explicitly added (opt-in) instead of
automatically being used (actually unconditionally, without an opt-out
option).

This has popped up in https://github.com/cda-tum/mqt-qcec/pull/396 since
`delocate` fails to properly repair the produced wheel. This is a result
of `gmp` being automatically installed on macOS runners via brew.
However, this gets the latest system version, which is not compatible
with the minimum macOS targets that we set. I believe this only popped
up now, because there has been a `delocate` update that catches this.
In lack of a better option at the moment, this PR makes GMP support
entirely opt-in, which disables it for wheel builds and resolves the
underlying issue.
Anyone that wants to build MQT Core with GMP support, can now configure
CMake with `-DMQT_CORE_WITH_GMP`

In the future, it might be worth investigating whether there is a better
way to do this, e.g., by manually building `gmp` with the required
compatibility.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

Signed-off-by: burgholzer <burgholzer@me.com>