🚸 make `dd/Package.hpp` export `dd/Package_fwd.hpp` (#640)
## Description

This PR slighlty improves the usability of the IWYU features of
clang-tidy for the DD package. To this end, it marks the
`dd/Package_fwd.hpp` include in the `dd/Package.hpp` header as an
export. This way, header need not include both headers, but can only
import the `dd/Package.hpp` header to also get the content of the foward
declaration header.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Signed-off-by: burgholzer <burgholzer@me.com>