🚸♻️ convenience methods and updates for DD functionality (#674)
## Description

This pull request includes several changes aimed at improving code
efficiency and readability across the DD package.

### Utility (functions):

* Introduced new utility functions in `Operations.hpp` for applying
unitary operations, measurements, resets, and classically controlled
operations on quantum states.
* DD routines for state preparation now automatically increase the
respective refcount, so this does not have to be done manually anymore.
* The classic-controlled Operation class now exposes its comparison kind
to the QuantumComputation class and to Python.

### Code simplification:

* Simplified the `getDD` function by removing redundant checks and
improving the handling of permutations.
* A `sample` method was introduced as a shortcut for simply sampling
from the output distribution of a circuit.
* Several tests were simplified and streamlined.

These changes collectively enhance the maintainability and performance
of the codebase.
They will require changes in consuming libraries. Especially mqt-ddsim,
mqt-qcec, and mqt-ddvis. However, these changes should simplify the
overall code in these libraries.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Signed-off-by: burgholzer <burgholzer@me.com>