🚸 slightly tweak how optional qiskit dependency is handled (#797)
## Description

This PR slightly adjusts how the optional qiskit dependency is handled
in the `load` method.
At that point in the function, the user is passing in a Qiskit
`QuantumCircuit` anyway, so Qiskit will naturally be available.

An observation I had while working on this:
We currently do not enforce a particular version constraint on qiskit at
runtime.
Thus, someone might use a version of qiskit that is too old at runtime,
and will probably run into errors.
We could probably do a better job of suggesting people to install the
`mqt.core[qiskit]` extra to guarantee compatibility.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

Signed-off-by: burgholzer <burgholzer@me.com>