🔒 update to latest `uv` version that properly supports dynamically versioned packages (#802)
## Description

The latest update to `uv` adds proper support for packages managed by
`setuptools-scm`.
Version information for packages with dynamic versioning is now not
explicitly included in the lock file.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

Signed-off-by: burgholzer <burgholzer@me.com>