Refactor handling of gate matrices and inverses (#752)
## Description

These commits change how matrices are created from an OpType that
reduces the number of switch statements necessary.
The downside of this approach is that some functions can be called with
arguments that do not affect the returned value. As the user never calls
these functions, I see this as acceptable.

I also removed the CX_MAT and CZ_MAT, as they are only used in tests and
do not occur in any other repository.

Fixes #484 

## Checklist:

- [x] The pull request only contains commits that are related to it.
- [ ] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.


I have not squashed the commits to make them easier to review, but I can
squash them if desired.

---------

Signed-off-by: burgholzer <burgholzer@me.com>
Co-authored-by: burgholzer <burgholzer@me.com>